Added transparency fix for xterms running tmux.
This is specifically xterms called running tmux immediately. Usually (such as
in a regular xterm) the transparency is loaded in my .bashrc. With tmux, the
xterm doesn't source my .bashrc, but just runs the tmux command first, so the
transparency doesn't happen. Tmux should be reading the transparency when *it*
goes through my .bashrc, but it apparently needs a bit of time before it can
load it. If I use a sleep .40 (as small as I've been able to get it), tmux will
set the transparency correctly when reading the .bashrc. Explicitly adding the
line of transparency before xterm opens tmux seems to work just fine though
(though quite hacky...).